Bug Fix - Group name in product attribute install script.
The group name changed from General to Product Detail, and when the install script ran it somehow (BUG) altered the first group and changed it's name, where I believe it should have created a new group if it didn't match an existing one. The problem then arose when saving the attribute set, it saw that the group name was different so it automatically changed the group code. The changing of the group code is was caused configurable products to not be editable any longer because whatever displays the details for configurable products looks for the product-detail tab in the product edit page to add the stuff for configurable products. When the configurable product data didn't exist on a product and it got saved it was interpreted as a simple product and changed as if someone had removed all the associated configurable items.